# api/openapi.json
{
  "components": {
    "schemas": {
      "ChatRequest": {
        "additionalProperties": false,
        "description": "`POST /v1/chat` body.",
        "properties": {
          "question": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "session": {
            "$ref": "#/components/schemas/EngineSession"
          }
        },
        "required": [
          "session",
          "question"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatResponse": {
        "additionalProperties": false,
        "description": "`POST /v1/chat` result.",
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationModel"
            },
            "title": "Citations",
            "type": "array"
          },
          "faithfulness": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Faithfulness"
          },
          "fallback": {
            "title": "Fallback",
            "type": "boolean"
          },
          "retrieved": {
            "title": "Retrieved",
            "type": "integer"
          },
          "standaloneQuery": {
            "title": "Standalonequery",
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        },
        "required": [
          "answer",
          "citations",
          "fallback",
          "standaloneQuery",
          "retrieved",
          "usage"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
      "ChatTurn": {
        "additionalProperties": false,
        "description": "A retained conversation turn, used for query compression.",
        "properties": {
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationModel"
            },
            "title": "Citations",
            "type": "array"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "createdAt": {
            "default": "",
            "title": "Createdat",
            "type": "string"
          },
          "fallback": {
            "default": false,
            "title": "Fallback",
            "type": "boolean"
          },
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "ChatTurn",
        "type": "object"
      },
      "CitationModel": {
        "additionalProperties": false,
        "description": "A retrieved source surfaced before the first token.",
        "properties": {
          "chunkId": {
            "title": "Chunkid",
            "type": "string"
          },
          "documentId": {
            "title": "Documentid",
            "type": "string"
          },
          "documentName": {
            "title": "Documentname",
            "type": "string"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "snippet": {
            "title": "Snippet",
            "type": "string"
          }
        },
        "required": [
          "chunkId",
          "documentId",
          "documentName",
          "page",
          "score",
          "snippet"
        ],
        "title": "CitationModel",
        "type": "object"
      },
      "DocumentSummary": {
        "additionalProperties": false,
        "description": "Indexed document summary returned to the browser.",
        "properties": {
          "chunkCount": {
            "title": "Chunkcount",
            "type": "integer"
          },
          "hasImages": {
            "title": "Hasimages",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "imageCount": {
            "title": "Imagecount",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "pageCount": {
            "title": "Pagecount",
            "type": "integer"
          },
          "sizeBytes": {
            "title": "Sizebytes",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "sizeBytes",
          "pageCount",
          "chunkCount",
          "imageCount",
          "hasImages"
        ],
        "title": "DocumentSummary",
        "type": "object"
      },
      "EngineDocument": {
        "additionalProperties": false,
        "description": "A PDF handed to the engine for indexing.",
        "properties": {
          "base64": {
            "minLength": 1,
            "title": "Base64",
            "type": "string"
          },
          "id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "pageCount": {
            "maximum": 5000.0,
            "minimum": 0.0,
            "title": "Pagecount",
            "type": "integer"
          },
          "sizeBytes": {
            "maximum": 2097152.0,
            "minimum": 1.0,
            "title": "Sizebytes",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "sizeBytes",
          "base64",
          "pageCount"
        ],
        "title": "EngineDocument",
        "type": "object"
      },
      "EngineSession": {
        "additionalProperties": false,
        "description": "Everything the engine needs to answer or evaluate for one session.",
        "properties": {
          "apiKey": {
            "default": "",
            "title": "Apikey",
            "type": "string"
          },
          "config": {
            "$ref": "#/components/schemas/PipelineConfig"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/EngineDocument"
            },
            "maxItems": 3,
            "title": "Documents",
            "type": "array"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ChatTurn"
            },
            "maxItems": 200,
            "title": "History",
            "type": "array"
          },
          "sessionId": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Sessionid",
            "type": "string"
          }
        },
        "required": [
          "sessionId",
          "config"
        ],
        "title": "EngineSession",
        "type": "object"
      },
      "EvaluateRequest": {
        "additionalProperties": false,
        "description": "`POST /v1/evaluate` body.",
        "properties": {
          "sampleCount": {
            "default": 4,
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Samplecount",
            "type": "integer"
          },
          "session": {
            "$ref": "#/components/schemas/EngineSession"
          }
        },
        "required": [
          "session"
        ],
        "title": "EvaluateRequest",
        "type": "object"
      },
      "EvaluationReport": {
        "additionalProperties": false,
        "description": "`POST /v1/evaluate` result.",
        "properties": {
          "category": {
            "const": "Retrieval Augmented Generation",
            "default": "Retrieval Augmented Generation",
            "title": "Category",
            "type": "string"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "documentCount": {
            "title": "Documentcount",
            "type": "integer"
          },
          "durationMs": {
            "title": "Durationms",
            "type": "integer"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/MetricResult"
            },
            "title": "Metrics",
            "type": "array"
          },
          "sampleCount": {
            "title": "Samplecount",
            "type": "integer"
          },
          "samples": {
            "items": {
              "$ref": "#/components/schemas/EvaluationSample"
            },
            "title": "Samples",
            "type": "array"
          }
        },
        "required": [
          "createdAt",
          "durationMs",
          "sampleCount",
          "documentCount",
          "metrics",
          "samples"
        ],
        "title": "EvaluationReport",
        "type": "object"
      },
      "EvaluationSample": {
        "additionalProperties": false,
        "description": "A single evaluated question, retained for inspection.",
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationModel"
            },
            "title": "Citations",
            "type": "array"
          },
          "contexts": {
            "items": {
              "type": "string"
            },
            "title": "Contexts",
            "type": "array"
          },
          "fallback": {
            "title": "Fallback",
            "type": "boolean"
          },
          "groundTruth": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Groundtruth"
          },
          "question": {
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question",
          "answer",
          "contexts",
          "citations",
          "fallback"
        ],
        "title": "EvaluationSample",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MetricResult": {
        "additionalProperties": false,
        "description": "One Ragas-style metric score.",
        "properties": {
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "samples": {
            "title": "Samples",
            "type": "integer"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "const": "N/A",
                "type": "string"
              }
            ],
            "title": "Score"
          },
          "skippedReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skippedreason"
          }
        },
        "required": [
          "metric",
          "score",
          "samples",
          "reason"
        ],
        "title": "MetricResult",
        "type": "object"
      },
      "PipelineConfig": {
        "additionalProperties": false,
        "description": "Immutable pipeline configuration, validated on arrival.",
        "properties": {
          "baseUrl": {
            "title": "Baseurl",
            "type": "string"
          },
          "chunkOverlapPercent": {
            "maximum": 20.0,
            "minimum": 10.0,
            "title": "Chunkoverlappercent",
            "type": "integer"
          },
          "chunkOverlapTokens": {
            "minimum": 0.0,
            "title": "Chunkoverlaptokens",
            "type": "integer"
          },
          "chunkSize": {
            "maximum": 2048.0,
            "minimum": 128.0,
            "multipleOf": 32.0,
            "title": "Chunksize",
            "type": "integer"
          },
          "distanceMetric": {
            "enum": [
              "cosine",
              "dot",
              "euclidean"
            ],
            "title": "Distancemetric",
            "type": "string"
          },
          "embeddingDimension": {
            "maximum": 8192.0,
            "minimum": 1.0,
            "title": "Embeddingdimension",
            "type": "integer"
          },
          "embeddingModel": {
            "enum": [
              "text-embedding-3-small",
              "text-embedding-3-large",
              "nomic-embed-text",
              "bge-m3",
              "embeddinggemma",
              "mxbai-embed-large"
            ],
            "title": "Embeddingmodel",
            "type": "string"
          },
          "maxInputTokens": {
            "maximum": 4096.0,
            "minimum": 256.0,
            "multipleOf": 32.0,
            "title": "Maxinputtokens",
            "type": "integer"
          },
          "model": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "enum": [
              "openai",
              "vocareum",
              "deepseek",
              "ollama"
            ],
            "title": "Provider",
            "type": "string"
          },
          "retrievalMode": {
            "enum": [
              "context-injection",
              "agentic"
            ],
            "title": "Retrievalmode",
            "type": "string"
          },
          "topK": {
            "maximum": 10.0,
            "minimum": 3.0,
            "title": "Topk",
            "type": "integer"
          }
        },
        "required": [
          "provider",
          "baseUrl",
          "model",
          "embeddingModel",
          "embeddingDimension",
          "chunkSize",
          "chunkOverlapPercent",
          "chunkOverlapTokens",
          "maxInputTokens",
          "distanceMetric",
          "topK",
          "retrievalMode"
        ],
        "title": "PipelineConfig",
        "type": "object"
      },
      "ResetResponse": {
        "additionalProperties": false,
        "description": "`DELETE /v1/session/{id}` result.",
        "properties": {
          "cleared": {
            "title": "Cleared",
            "type": "boolean"
          }
        },
        "required": [
          "cleared"
        ],
        "title": "ResetResponse",
        "type": "object"
      },
      "TestConnectionRequest": {
        "additionalProperties": false,
        "description": "`POST /v1/test` body.",
        "properties": {
          "session": {
            "$ref": "#/components/schemas/EngineSession"
          }
        },
        "required": [
          "session"
        ],
        "title": "TestConnectionRequest",
        "type": "object"
      },
      "TestConnectionResponse": {
        "additionalProperties": false,
        "description": "`POST /v1/test` result.",
        "properties": {
          "embeddingDimension": {
            "title": "Embeddingdimension",
            "type": "integer"
          },
          "embeddingProbe": {
            "title": "Embeddingprobe",
            "type": "boolean"
          },
          "latencyMs": {
            "title": "Latencyms",
            "type": "integer"
          },
          "modelEcho": {
            "title": "Modelecho",
            "type": "string"
          },
          "reachable": {
            "title": "Reachable",
            "type": "boolean"
          }
        },
        "required": [
          "reachable",
          "modelEcho",
          "latencyMs",
          "embeddingDimension",
          "embeddingProbe"
        ],
        "title": "TestConnectionResponse",
        "type": "object"
      },
      "UpsertResponse": {
        "additionalProperties": false,
        "description": "`POST /v1/pipeline` result.",
        "properties": {
          "chunkCount": {
            "title": "Chunkcount",
            "type": "integer"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationModel"
            },
            "title": "Citations",
            "type": "array"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentSummary"
            },
            "title": "Documents",
            "type": "array"
          },
          "engineSessionId": {
            "title": "Enginesessionid",
            "type": "string"
          },
          "multimodal": {
            "title": "Multimodal",
            "type": "boolean"
          }
        },
        "required": [
          "engineSessionId",
          "documents",
          "chunkCount",
          "citations",
          "multimodal"
        ],
        "title": "UpsertResponse",
        "type": "object"
      },
      "Usage": {
        "additionalProperties": false,
        "description": "Token accounting for cost attribution.",
        "properties": {
          "completionTokens": {
            "default": 0,
            "title": "Completiontokens",
            "type": "integer"
          },
          "promptTokens": {
            "default": 0,
            "title": "Prompttokens",
            "type": "integer"
          }
        },
        "title": "Usage",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Retrieval engine for RAGdoll: PDF ingest, chunking, embedding, retrieval, grounded generation and Ragas-style evaluation.",
    "title": "RAGdoll RAG Engine",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/chat": {
      "post": {
        "description": "Answers a question with citations and the groundedness gate applied.",
        "operationId": "chat_v1_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat",
        "tags": [
          "chat"
        ]
      }
    },
    "/v1/chat/stream": {
      "post": {
        "description": "Streams the answer as SSE: citations, status, tokens, then done.",
        "operationId": "chat_stream_v1_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat Stream",
        "tags": [
          "chat"
        ]
      }
    },
    "/v1/evaluate": {
      "post": {
        "description": "Runs the Ragas-style metric suite.",
        "operationId": "evaluate_v1_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate",
        "tags": [
          "evaluate"
        ]
      }
    },
    "/v1/pipeline": {
      "post": {
        "description": "Builds (or replaces) the session index from the uploaded PDFs.",
        "operationId": "upsert_pipeline_v1_pipeline_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpsertResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upsert Pipeline",
        "tags": [
          "pipeline"
        ]
      }
    },
    "/v1/session/{session_id}": {
      "delete": {
        "description": "Purges the index, documents and history for a session.",
        "operationId": "reset_session_v1_session__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Session",
        "tags": [
          "pipeline"
        ]
      }
    },
    "/v1/session/{session_id}/citations": {
      "get": {
        "description": "Returns the citation catalogue for an indexed session.",
        "operationId": "session_citations_v1_session__session_id__citations_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Session Citations V1 Session  Session Id  Citations Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Session Citations",
        "tags": [
          "chat"
        ]
      }
    },
    "/v1/test": {
      "post": {
        "description": "Validates the provider, key, model and embedding width.",
        "operationId": "test_connection_v1_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestConnectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Connection",
        "tags": [
          "pipeline"
        ]
      }
    }
  },
...
}